pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-smart-control-system"
version = "0.1.0"
description = "Smart Traffic Control RL Environment for OpenEnv"
requires-python = ">=3.10"

dependencies = [
    "openenv-core[core]>=0.2.2",
    "matplotlib",
    "gradio>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0"
]